+++ lib/src/models/models.dart
@@ -5,4 +5,5 @@ export 'redirectly_config.dart';
 export 'redirectly_error.dart';
 export 'redirectly_link.dart';
 export 'redirectly_link_click.dart';
+export 'redirectly_link_resolution.dart';
 export 'redirectly_temp_link.dart';